+++ Runtime/CSharp/HumrRecordingLoader.cs
@@ -15,7 +15,6 @@ namespace DrSakuu.Humr
         public int fileIndex;
         public int targetIndex;
         public bool exportFbx = true;
-        public bool exportAnim;
         public bool showAdvanced;
         public bool blenderHipFix = true;
         public Animator Animator => GetComponent<Animator>();