+++ Documentation/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Toggle for world absolute position recording
 - Button to refresh Log Files and reset Log Path
+- Buttons to export individual takes as .anim
 
 ### Fixed